third_party/blink/web_tests/fast/forms/select/listbox-deselect-scroll-expected.txt

This tests that deselecting an option won't cause unnecessary scrolling.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS sl.scrollTop is scrollBeforeClick
PASS selectionPattern(sl) is "11111110111"
PASS successfullyParsed is true

TEST COMPLETE

Accessibility
CSS
Drosera
Evangelism
Forms
Frames
History
HTML DOM
HTML Editing
Images
SCROLL UP
